# Fix: Make no-multi-spaces work for every case (fixes #1603, fixes #1659)