--- a/lib/eslint.js
+++ b/lib/eslint.js
@@ -370,7 +370,8 @@ module.exports = (function() {
         controller = null,
         reportingConfig = [],
         commentLocsEnter = [],
-        commentLocsExit = [];
+        commentLocsExit = [],
+        currentAST = null;
 
     /**
      * Parses text into an AST. Moved out here because the try-catch prevents
@@ -495,6 +496,7 @@ module.exports = (function() {
     api.reset = function() {
         this.removeAllListeners();
         messages = [];
+        currentAST = null;
         currentConfig = null;
         currentText = null;
         currentTextLines = [];
@@ -537,6 +539,9 @@ module.exports = (function() {
 
         // if espree failed to parse the file, there's no sense in setting up rules
         if (ast) {
+
+            currentAST = ast;
+
             // process initial config to make it safe to extend
             config = prepareConfig(config);
 
@@ -736,6 +741,14 @@ module.exports = (function() {
      */
     api.getSourceLines = function() {
         return currentTextLines;
+    };
+
+    /**
+     * Retrieves an array containing all comments in the source code.
+     * @returns {ASTNode[]} An array of comment nodes.
+     */
+    api.getAllComments = function() {
+        return currentAST.comments;
     };
 
     /**