--- a/lib/rule-context.js
+++ b/lib/rule-context.js
@@ -25,7 +25,9 @@ var PASSTHROUGHS = [
         "getAncestors",
         "getScope",
         "getJSDocComment",
-        "getFilename"
+        "getFilename",
+        "getTokenByRangeIndex",
+        "getAllComments"
     ];
 
 //------------------------------------------------------------------------------