--- a/lib/token-store.js
+++ b/lib/token-store.js
@@ -14,6 +14,7 @@ module.exports = function(tokens) {
     var api = {},
         starts = Object.create(null),
         ends = Object.create(null),
+        ranges = Object.create(null),
         index, length, range;
 
     /**
@@ -64,6 +65,11 @@ module.exports = function(tokens) {
         range = tokens[index].range;
         starts[range[0]] = index;
         ends[range[1]] = index;
+
+        // allows fast lookup of any source index
+        for (var j = range[0]; j < range[1]; j++) {
+            ranges[j] = index;
+        }
     }
 
     /**
@@ -180,6 +186,15 @@ module.exports = function(tokens) {
             lastTokenIndex(left) + 1 - padding,
             starts[right.range[0]] + padding
         );
+    };
+
+    /**
+     * Returns the token within with the given source index occurs.
+     * @param {int} rangeIndex The index at which to look for the token.
+     * @returns {Token} The token at that position or null if not found.
+     */
+    api.getTokenByRangeIndex = function(rangeIndex) {
+        return tokens[ranges[rangeIndex]] || null;
     };
 
     return api;