This was an absolute monster to work through and it's still not quite there. One edge case remaining. Just posting to share.

I had to create a map of where all tokens are, and that subsequently slowed things down a bit. I think the slowdown is worth it for the tradeoff.

$ npm run perf

> eslint@0.12.0 perf c:\Users\Nicholas\projects\personal\jscheck
> node Makefile.js perf

```
# Before

CPU Speed is 3093 with multiplier 7500000
Performance Run #1:  1501.513292ms
Performance Run #2:  1516.6144359999998ms
Performance Run #3:  1522.233913ms
Performance Run #4:  1537.468143ms
Performance Run #5:  1510.7205079999999ms
Performance budget ok:  1516.6144359999998ms (limit: 2424.8302618816683ms)

# After

CPU Speed is 3093 with multiplier 7500000
Performance Run #1:  1573.712675ms
Performance Run #2:  1545.923181ms
Performance Run #3:  1556.752278ms
Performance Run #4:  1558.050048ms
Performance Run #5:  1550.235287ms
Performance budget ok:  1556.752278ms (limit: 2424.8302618816683ms)
```