@gcochard 2015-01-24
This is exactly what I was thinking was needed, but I didn't have the time necessary for such a large refactor. Looks like a great start!