@btmills 2015-01-27
Philosophical question, not about the code: since this should conflict with key-spacing in its "align" mode, if somebody asks, is the answer that these rules have become mutually exclusive? I'm not advocating exceptions here - that sounds excessively complex.

I have one concern in the code about an upper bound. Other than that, LGTM.

No need to hold this PR until I can investigate a range tree - I think the odds are low it'll be anything but slower. I'll file an issue if it turns out to be an improvement.