@nzakas 2015-01-27
Good question . We've always shied away from mutually exclusive rules because they are confusing. That said, I'm not sure how this rule could be made aware of exceptions in other rules. :-/