@btmills 2015-01-27
Since mutual exclusivity is undesirable, I'll throw out an idea. What if there were a list of excepted node types, such that if the most recent ancestor containing a whitespace range is in the excepted list, it doesn't report? For example:

```js
({
	a: b
})
```
If `Property` were excepted, the space between the colon and the property value would be contained within the `Property` node, but not the key or value.

This is just a brainstorm. No idea how practical it is. But I think that would solve key-spacing at least.