@nzakas 2015-01-28
We're not really dealing with the AST in this rule, it's just strings and tokens. We'd need a way to match a range index with an AST node.