@btmills 2015-01-28
> We'd need a way to match a range index with an AST node.

Rather than throwing a bunch of pseudocode here, I built such a way in #1738 :smile: