@btmills 2015-01-30
Aha! The ranges aren't necessary! I think. Does this logic hold?

A space sequence can end one of five ways:

1. It reaches the end of the line.
This rule is not designed to handle end-of-line spaces. Nothing should be reported.

1. It reaches the start of a comment.
The `isIndexInComment` check will find that `pattern.lastIndex` is the starting index of the comment. Nothing will be reported.

1. The space sequence is within a comment.
Again, `isIndexInComment` will prevent reporting.

1. The space sequence is within a string or template literal.
Nothing should be reported.

1. It encounters the start of a token.
In this case, `pattern.lastIndex` will be exactly the starting index of some token. The rule should report an error.

If I'm not missing anything, then the rule should only report when it encounters a token starting at `pattern.lastIndex`. Therefore, we replace `getTokenByRangeIndex` with a new method, `getTokenStartingAtIndex`, consisting simply of `return tokens[starts[index]] || null;`. The rule warns if and only if `getTokenStartingAtIndex(pattern.lastIndex)` returns a token. No more retrieval from the middle of a token's range, and no more need to assemble the map of ranges.

[I implemented exactly this strategy](https://github.com/eslint/eslint/commit/7fedd81fd087548d0a72be24da4d16eb98e6c0fc) on top of this PR and pushed it to [a new branch](https://github.com/eslint/eslint/tree/getTokenStartingAtIndex). The existing tests all pass. Can you think of any new ones that might cause it to fail?