@nzakas 2015-01-30
Nice, that's much simpler. Want to open a PR?